+++ frontend/src/components/stats/BullsBearProgress.jsx
