+++ frontend/src/components/stats/CPRStats.jsx
